requirements.txt
absl-py==0.10.0
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
blinker==1.4
Brlapi==0.6.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.12
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu4
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.11.3
idna==2.6
importlib-metadata==2.0.0
introcs==1.2.2
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.1.1
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.10.0
Magnus==1.0.1
Markdown==3.2.2
MarkupSafe==1.1.1
mate-hud==19.10.0
mate-menu==19.10.2
mate-tweak==19.10.4
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.2
oauthlib==2.1.0
olefile==0.46
onboard==1.4.1
opencv-contrib-python==4.2.0.34
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.5.1
ptyprocess==0.6.0
pulsemixer==1.5.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygame==2.0.0
Pygments==2.7.1
PyGObject==3.34.0
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.9.0+ubuntu1.4
python-dateutil==2.7.3
python-debian==0.1.36
python-registry==1.3.1
python-xapp==1.6.0
python-xlib==0.23
pytz==2020.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.12.0
sklearn==0.0
systemd-python==234
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==2.0.0b1
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.24.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
yarp==1.0.28
zipp==3.2.0
